Control.Monad.Codec

Documentation

data CodecFor r w c a

A serializer/deserializer pair reading a in context r and writing c in context w.

Constructors

Codec

Fields

codecIn :: r a

codecOut :: c -> w a

type Codec r w a = CodecFor r w a a

(=.) :: (c' -> c) -> CodecFor r w c a -> CodecFor r w c' a

Compose a function into the serializer of a Codec. Useful to modify a Codec so that it writes a particular record field.

fmapArg :: Functor f => (a -> f ()) -> a -> f a

Modify a serializer function so that it also returns the serialized value, Useful for implementing codecs.