Data.Codec.TH

Documentation

genFields :: Name -> Q [Dec]

Generate Fields for a given data type. Currently only single-constructor records are supported. Each record field a will be turned into a Field f_a, and all constructors will be turned into Cons.