Data.Aeson.Codec

Documentation

data JSONCodec a

Describes the de/serialization of a type a. Equivalent to a ToJSON and a FromJSON instance.

Constructors

JSONCodec

Fields

parseJSONCodec :: Value -> Parser a

toJSONCodec :: a -> Value

toEncodingCodec :: a -> Encoding

defJSON :: (FromJSON a, ToJSON a) => JSONCodec a

Encode/decode a value with its ToJSON and FromJSON instances.

JSON object codecs

type ObjectParser = ReaderT Object Parser

type ObjectBuilder = Writer (Series, Endo [Pair])

type ObjectCodec a = Codec ObjectParser ObjectBuilder a

A codec that parses values out of a given Object, and produces key-value pairs into a new one.

field :: (FromJSON a, ToJSON a) => Text -> ObjectCodec a

Store/retrieve a value in a given JSON field, with the default JSON serialization.

field' :: Text -> JSONCodec a -> ObjectCodec a

Store/retrieve a value in a given JSON field, with a given JSONCodec.

asObject :: String -> ObjectCodec a -> JSONCodec a

Turn an ObjectCodec into a JSONCodec with an expected name (see withObject).

JSON array codecs

type ArrayParser = StateT [Value] Parser

type ArrayBuilder = Writer (Series, [Value])

type ArrayCodec a = Codec ArrayParser ArrayBuilder a

A codec that serializes data to a sequence of JSON array elements.

element :: (FromJSON a, ToJSON a) => ArrayCodec a

Expect/append an array element, using the default serialization.

element' :: JSONCodec a -> ArrayCodec a

Expect/append an array element, using a given JSONCodec.

asArray :: String -> ArrayCodec a -> JSONCodec a

A codec that parses values out of a given Array, and produces key-value pairs into a new one.

arrayOf :: (FromJSON b, ToJSON b) => (a -> [b]) -> ([b] -> a) -> JSONCodec a

Given a a way to turn a into [ b ] and back, create a JSONCodec for a.

arrayOf' :: (a -> [b]) -> ([b] -> a) -> JSONCodec b -> JSONCodec a

Given a JSONCodec for b and a way to turn a into [ b ] and back, create a JSONCodec for a.